# api/internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "UI起点・初心者向け OSS DAST（Nuclei ラッパー）の API。スキャン受付・サイト管理・ドメイン所有確認・認証情報の暗号化保管・ダッシュボード/レポート用データ提供。\n認証情報の生値はレスポンスに一切含めない（マスク）。スキャン実行は worker プロセスに分離（ADR-0001）。",
        "title": "goodast API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/sites": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "サイト一覧を取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_handler.siteResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "サイト名とベース URL を登録。非ローカルは所有確認トークン＋設置ガイドを返す。localhost 等は即 verified。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "サイトを登録",
                "parameters": [
                    {
                        "description": "サイト登録",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.registerSiteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.siteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sites/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "サイト詳細を取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.siteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/sites/{id}/verify": {
            "post": {
                "description": "設置済みトークンを検証する。未達（未設置・不一致）は 422。成功で ownership_verified=true。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "ドメイン所有確認を実行",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.siteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid site id"
                }
            }
        },
...
        "internal_handler.registerSiteRequest": {
            "type": "object",
            "required": [
                "base_url",
                "name"
            ],
            "properties": {
                "base_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "verify_method": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.siteResponse": {
            "type": "object",
            "properties": {
                "base_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ownership_verified": {
                    "type": "boolean"
                },
                "verification": {
                    "$ref": "#/definitions/internal_handler.verificationGuide"
                },
                "verify_method": {
                    "type": "string"
                },
                "verify_token": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.verificationGuide": {
            "type": "object",
            "properties": {
                "dns_record": {
                    "type": "string"
                },
                "file_content": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                }
            }
        }
    }
}